fix: DetailScreenクラスは必要ない
